requirements.txt
streamlit==1.28.0
requests==2.31.0
python-dotenv==1.0.0
pdfplumber==0.10.0
python-docx>=1.0.0
groq>=0.31.1
